# Conflicts:
#	Backend/PropNetApp/pom.xml
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/city/City.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/features/email/EmailDetail.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/features/email/EmailDetailRestController.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/features/email/EmailDetailServiceImpl.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/governorate/Governorate.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/locality/Locality.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/localitydetail/LocalityDetail.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/post/Post.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/post/PostRestController.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/post/PostServiceImpl.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/realestatellisting/IRealEstateListingService.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/realestatellisting/RealEstateListing.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/realestatellisting/RealEstateListingRepository.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/realestatellisting/RealEstateListingRestController.java
#	Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/realestatellisting/RealEstateListingServiceImpl.java
#	Backend/PropNetApp/src/main/resources/application.properties
#	Frontend/src/app/administration/Annonce/list-annonce/list-annonce.component.ts
#	Frontend/src/app/administration/administration-routing.module.ts
#	Frontend/src/app/administration/dashboard/dashboard.component.ts
#	Frontend/src/app/administration/forum/list-posts/list-posts.component.ts
#	Frontend/src/app/app.module.ts
#	Frontend/src/app/detail-annonce/detail-annonce.component.ts
#	Frontend/src/app/services/post.service.ts

+++ Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/governorate/GovernorateRestController.java
@@ -20,11 +20,11 @@ public class GovernorateRestController {
         return governorateService.retrieveAllGovernorates();
     }
     @GetMapping("/retrieve-governorate/{governorate-id}")
-    public Governorate retrieveGovernorate(@PathVariable("governorate-id") Long governorateId) {
+    public Governorate retrieveGovernorate(@PathVariable("governorate-id") Integer governorateId) {
         return governorateService.retrieveGovernorate(governorateId);
     }
     @DeleteMapping("/remove-governorate/{governorate-id}")
-    public void removeGovernorate(@PathVariable("governorate-id") Long governorateId) {
+    public void removeGovernorate(@PathVariable("governorate-id") Integer governorateId) {
         governorateService.deleteGovernorate(governorateId);
     }
     @PutMapping("/update-governorate")
