+++ Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/governorate/IGovernorateService.java
@@ -7,6 +7,6 @@ public interface IGovernorateService {
     Governorate addGovernorate(Governorate governorate);
     List<Governorate> retrieveAllGovernorates();
     Governorate updateGovernorate(Governorate governorate);
-    Governorate retrieveGovernorate(Long idGovernorate);
-    void deleteGovernorate(Long idGovernorate);
+    Governorate retrieveGovernorate(Integer idGovernorate);
+    void deleteGovernorate(Integer idGovernorate);
 }
